RestService/requirements.txt
aniso8601==6.0.0
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
pluggy==0.9.0
py==1.8.0
PyHamcrest==1.9.0
pytz==2018.9
six==1.12.0
Werkzeug==0.15.2
